fixed login button covering the  cfc title and home-view hero page (#102)
* fixed login button covering the whole cfc title and home-view hero page length

* revision 1: Add login to drawer
